requirements.txt
ccxt
pandas
clickhouse-connect
plotly
python-dotenv
flask
